todo_phase2/Full-Stack-Web-Application-hackathon2-phase2/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
fastapi-cors
python-multipart
bcrypt==4.0.1